requirement.txt
asyncio
python-dotenv
aiohttp
beautifulsoup4
prisma